monitor/requirements.txt
python-dotenv==1.0.1

celery==5.4.0
redis==5.2.1
flower==2.0.1
